+++ main.py
@@ -5,7 +5,7 @@ main.py: game launcher
 import ffmCLI
 import os
 
-version = "0.5.0_wip"
+version = "0.6.0"
 
 user_name = input("What is your name? ")
 user_name_with_underscores = user_name.replace(" ", "_")