+++ players/active_player.py
@@ -1,7 +1,0 @@
-# TODO everything
-
-class ActivePlayer:
-  def __init__(self, name, **kwargs):
-    pass
-
-# TODO make a fifa loader first