+++ players/fifa.py
@@ -1,1 +1,0 @@
-# thismwill be the fifal loader